requirements.txt
python-telegram-bot>=22.5
python-dotenv>=1.2.0
psycopg2-binary>=2.9.9

